feat: Export CLI as part of library
