+++ rasn-compiler/src/lib.rs
@@ -7,6 +7,9 @@ mod lexer;
 mod tests;
 mod validator;
 
+#[cfg(feature = "cli")]
+pub mod cli;
+
 use std::{
     cell::RefCell,
     collections::BTreeMap,
